requirements.txt
python-dotenv
python-binance
